move web storybook to example
