+++ examples/ai-lab-example/.storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+  stories: ['../**/*.stories.mdx', '../**/*.stories.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
+};
