+++ examples/ai-lab-example/tsconfig.json
@@ -8,7 +8,7 @@
     // output .js.map sourcemap files for consumers
     "sourceMap": true,
     // stricter type-checking for stronger correctness. Recommended by TS
-    "strict": true,
+    // "strict": true,
     // linter checks for common issues
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
