--- packages/ai-lab/.storybook/main.js
@@ -1,7 +1,0 @@
-module.exports = {
-  stories: [
-    '../components/**/*.stories.mdx',
-    '../components/**/*.stories.@(js|jsx|ts|tsx)',
-  ],
-  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
-};
