+++ packages/ai-lab/components/AILabImage/labels.ts
