+++ packages/ai-lab/package.json
@@ -14,21 +14,16 @@
     "start": "tsdx watch --verbose --noClean",
     "build": "tsdx build",
     "lint": "tsdx lint",
-    "prepublish": "npm run build",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "prepublish": "npm run build"
   },
   "author": "Gant Laborde",
   "license": "MIT",
   "devDependencies": {
     "@babel/core": "^7.15.5",
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-essentials": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/addon-storyshots": "^6.3.12",
-    "@storybook/react": "^6.3.12",
+    "@types/core-js": "^2.5.5",
     "@types/react": "^17.0.24",
     "@types/react-dom": "^17.0.9",
+    "@types/regenerator-runtime": "^0.13.1",
     "babel-loader": "^8.2.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -37,8 +32,6 @@
   "dependencies": {
     "@tensorflow/tfjs": "2.6.0",
     "@tensorflow/tfjs-node": "^3.0.0",
-    "@types/core-js": "^2.5.5",
-    "@types/regenerator-runtime": "^0.13.1",
     "core-js": "^3.8.0",
     "regenerator-runtime": "^0.13.7"
   }
